Refs #31349. Move general presenter functions to basic presenter.
