+++ scripts/Muon/GUI/Common/fitting_widgets/general_fitting/general_fitting_presenter.py
@@ -21,9 +21,6 @@ class GeneralFittingPresenter(BasicFittingPresenter):
         """Initialize the GeneralFittingPresenter. Sets up the slots and event observers."""
         super(GeneralFittingPresenter, self).__init__(view, model)
 
-        # This prevents plotting the wrong data when selecting different group/pairs on the grouping tab
-        self._update_plot = True
-
         self.fitting_mode_changed_notifier = GenericObservable()
         self.simultaneous_fit_by_specifier_changed = GenericObservable()
 
@@ -34,7 +31,6 @@ class GeneralFittingPresenter(BasicFittingPresenter):
         self.double_pulse_observer = GenericObserverWithArgPassing(self.handle_pulse_type_changed)
         self.model.context.gui_context.add_non_calc_subscriber(self.double_pulse_observer)
 
-        self.view.set_slot_for_dataset_changed(self.handle_dataset_name_changed)
         self.view.set_slot_for_fitting_mode_changed(self.handle_fitting_mode_changed)
         self.view.set_slot_for_simultaneous_fit_by_changed(self.handle_simultaneous_fit_by_changed)
         self.view.set_slot_for_simultaneous_fit_by_specifier_changed(self.handle_simultaneous_fit_by_specifier_changed)
@@ -68,17 +64,6 @@ class GeneralFittingPresenter(BasicFittingPresenter):
         self._update_plot = False
         self.update_and_reset_all_data()
         self._update_plot = True
-
-    def handle_fitting_finished(self, fit_function, fit_status, chi_squared) -> None:
-        """Handle when fitting is finished."""
-        self.update_fit_statuses_and_chi_squared_in_model(fit_status, chi_squared)
-        self.update_fit_function_in_model(fit_function)
-
-        self.update_fit_statuses_and_chi_squared_in_view_from_model()
-        self.update_fit_function_in_view_from_model()
-
-        self.selected_fit_results_changed.notify_subscribers(self.model.get_active_fit_results())
-        self.fit_parameter_changed_notifier.notify_subscribers()
 
     def handle_fitting_mode_changed(self) -> None:
         """Handle when the fitting mode is changed to or from simultaneous fitting."""
@@ -117,23 +102,6 @@ class GeneralFittingPresenter(BasicFittingPresenter):
 
         self.simultaneous_fit_by_specifier_changed.notify_subscribers()
 
-    def handle_dataset_name_changed(self) -> None:
-        """Handle when the display workspace combo box is changed."""
-        self.model.current_dataset_index = self.view.current_dataset_index
-
-        self.update_fit_statuses_and_chi_squared_in_view_from_model()
-        self.update_fit_function_in_view_from_model()
-        self.update_start_and_end_x_in_view_from_model()
-
-        if self._update_plot:
-            self.selected_fit_results_changed.notify_subscribers(self.model.get_active_fit_results())
-            self.model.update_plot_guess(self.view.plot_guess)
-
-    def set_selected_dataset(self, dataset_name: str) -> None:
-        """Sets the workspace to be displayed in the view programmatically."""
-        # Triggers handle_dataset_name_changed which updates the model
-        self.view.current_dataset_name = dataset_name
-
     def switch_fitting_mode_in_view(self) -> None:
         """Switches the fitting mode by updating the relevant labels and checkboxes in the view."""
         if self.model.simultaneous_fitting_mode:
@@ -152,25 +120,18 @@ class GeneralFittingPresenter(BasicFittingPresenter):
             self.model.fit_statuses = [fit_status] * self.model.number_of_datasets
             self.model.chi_squared = [chi_squared] * self.model.number_of_datasets
         else:
-            self.model.current_fit_status = fit_status
-            self.model.current_chi_squared = chi_squared
+            super().update_fit_statuses_and_chi_squared_in_model(fit_status, chi_squared)
 
     def update_fit_function_in_model(self, fit_function: IFunction) -> None:
         """Updates the fit function stored in the model. This is used after a fit."""
         if self.model.simultaneous_fitting_mode:
             self.model.simultaneous_fit_function = fit_function
         else:
-            self.model.current_single_fit_function = fit_function
+            super().update_fit_function_in_model(fit_function)
 
     def update_simultaneous_fit_by_specifiers_in_view(self) -> None:
         """Updates the entries in the simultaneous fit by specifier combo box."""
         self.view.setup_fit_by_specifier(self.model.get_simultaneous_fit_by_specifiers_to_display_from_context())
-
-    def update_dataset_names_in_view_and_model(self) -> None:
-        """Updates the datasets currently displayed. The simultaneous fit by specifier must be updated before this."""
-        super().update_dataset_names_in_view_and_model()
-        self.view.update_dataset_name_combo_box(self.model.dataset_names)
-        self.model.current_dataset_index = self.view.current_dataset_index
 
     def update_fit_functions_in_model_from_view(self) -> None:
         """Updates the fit functions stored in the model using the view."""
